+++ demo/components/dropdown/dropdown.component.ts
@@ -27,16 +27,26 @@ export class DropdownComponent implements OnInit {
     { label: 'item 10' }
   ];
 
-  onSelect(selectedItems: any[]) {
-    // console.log(selectedItems);
-    if (selectedItems && selectedItems[0]) {
-      this.selectedItem = selectedItems[0];
-    } else {
-      this.selectedItem = null;
-    }
+  items2: any[] = [
+    { label: 'item 1' },
+    { label: 'item 2' },
+    { label: 'item 3', class: 'myclass myclass2' },
+    { label: 'item 4 (preselect)', selected: true },
+    { label: 'item 5' },
+    { label: 'item 6', sublabel: 'sublabel of item 6' },
+    { label: 'item 7', sublabel: 'sublabel of item 7' },
+    { label: 'item 8', sublabel: 'sublabel of item 8' },
+    { label: 'item 9' },
+    { label: 'item 10' }
+  ];
+
+
+  onSelect(value: any) {
+    console.log('new dropdown-value:');
+    console.dir(value);
   }
 
-  expand1() {
+  expand() {
     this.expanded = !this.expanded;
   }
   expand2() {