+++ src/components/dropdown/dropdown.component.ts
@@ -4,7 +4,8 @@ import {
   Output,
   ChangeDetectionStrategy,
   EventEmitter,
-  forwardRef
+  forwardRef,
+  OnInit
 } from '@angular/core';
 import { ControlValueAccessor, NgControl, NG_VALUE_ACCESSOR } from '@angular/forms';
 
@@ -20,55 +21,70 @@ export const CUSTOM_INPUT_CONTROL_VALUE_ACCESSOR: any = {
   changeDetection: ChangeDetectionStrategy.OnPush,
   providers: [CUSTOM_INPUT_CONTROL_VALUE_ACCESSOR]
 })
-export class DropdownComponent implements ControlValueAccessor {
+export class DropdownComponent implements ControlValueAccessor, OnInit {
 
   @Input() items: any[];
-  @Input('value') value: string[] = []; // array of values of selected items
   @Input() tabindex: number = 0;
   @Input() expanded: boolean = false;
   @Input() maxSelectableItems: number = 1;
   @Input() minSelectableItems: number = 1;
   @Input() ariaRole: string = 'listbox';
 
+  get value(): any {
+    let ret = this.items
+      .filter(i => i.selected)
+      .map(i => i.value);
+    if (this.maxSelectableItems == 1) ret = ret[0];
+    return ret;
+  };
+  @Input() set value(v: any) {
+    if (!Array.isArray(v)) v = [v];
+    this.items
+      .forEach(i => {
+        if (v.includes(i.value)) i.selected = true;
+        else i.selected = false;
+      });
+    this.onChange();
+  }
+
   @Output('change') change$ = new EventEmitter<any[]>();
 
 
-  constructor() { }
+  ngOnInit() {
+    // ensure items have a value
+    this.items = this.items.map(i => {
+      if (!i.value) i.value = i.label;
+      if (!i.label) i.label = i.value;
+      return i;
+    });
+  }
+
+
+  selectedItems() {
+    return this.items.filter(i => i.selected);
+  }
 
   clickItem(item: any) {
     if (item.disabled) return;
 
-    // prevent overflow maxSelectableItems
-    if (
-      !item.selected &&
-      this.value.length >= this.maxSelectableItems
-    ) this.items.find(i => i.selected).selected = false;
-
-    if (
-      !item.selected &&
-      this.maxSelectableItems == 1
-    ) this.items.map(i => i.selected = false);
-
-    // prevent underflow minSelectableItems
-    if (
-      item.selected &&
-      this.value.length <= this.minSelectableItems
-    ) return;
-
+    if (!item.selected) {
+      // prevent overflow maxSelectableItems
+      if (this.selectedItems().length >= this.maxSelectableItems)
+        this.items.find(i => i.selected).selected = false;
+      if (this.maxSelectableItems == 1)
+        this.items.forEach(i => i.selected = false);
+    } else {
+      // prevent underflow minSelectableItems
+      if (this.selectedItems().length <= this.minSelectableItems) return;
+    }
     item.selected = !item.selected;
-    this.value = this.items
-      .filter(i => i.selected == true)
-      .map(i => i.value);
     this.onChange();
   }
 
-  onChange() {
-    let outValue: any = this.value;
-    if (this.maxSelectableItems <= 1)
-      outValue = this.value[0];
 
-    this.change$.emit(outValue);
-    !!this.onChangeCallback && this.onChangeCallback(outValue);
+  onChange() {
+    this.change$.emit(this.value);
+    !!this.onChangeCallback && this.onChangeCallback(this.value);
   }
 
   /**
@@ -77,14 +93,9 @@ export class DropdownComponent implements ControlValueAccessor {
   private onTouchedCallback: (_: any) => void;
   private onChangeCallback: (_: any) => void;
 
-  writeValue(value: any): void {
-    if (!Array.isArray(value)) value = [value];
-    this.value = value;
-    this.items
-      .forEach(i => {
-        if (value.includes(i.value)) i.selected = true;
-        else i.selected = false;
-      });
+  writeValue(v: any): void {
+    if (v.toString() != this.value.toStrig())
+      this.value = v;
   }
   registerOnChange(fn: any) {
     this.onChangeCallback = fn;