+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## master
 
 ### Breaking Changes
-- none
+- dropdown: see [readme](/#/dropdown)
 
 ## 15979eb (2016-12-14)
 
@@ -11,4 +11,3 @@
 - angular >= 2.3.0 required   
 - vcl-layer: The `<vcl-layer-base>` it added automatically to the app root   
 - wormhole: Renamed directives to `wormhole`/`connectWormhole`
-